code cleanup and add documentation
