+++ packages/data/src/index.js
@@ -9,13 +9,17 @@ import combineReducers from 'turbo-combine-reducers';
 import defaultRegistry from './default-registry';
 import * as plugins from './plugins';
 
-export { withSelect } from './components/with-select';
+export { default as withSelect } from './components/with-select';
 export { default as withDispatch } from './components/with-dispatch';
 export { default as withRegistry } from './components/with-registry';
-export { RegistryProvider, RegistryConsumer } from './components/registry-provider';
+export {
+	RegistryProvider,
+	RegistryConsumer,
+	useRegistry,
+} from './components/registry-provider';
+export { default as useSelect } from './components/use-select';
 export {
 	AsyncModeProvider as __experimentalAsyncModeProvider,
-	useAsyncMode as __experimentalUseAsyncMode,
 } from './components/async-mode-provider';
 export { createRegistry } from './registry';
 export { plugins };
